+++ .eslintrc.json
@@ -21,9 +21,7 @@
     ],
     "rules": {
         "prettier/prettier": ["error", {
-            
            "endOfLine":"auto"
-            
          }],
         "react/prop-types": 0
     }
