+++ .github/workflows/main.yml
@@ -2,7 +2,7 @@
 
 name: CI
 
-# Controls when the action will run. 
+# Controls when the action will run.
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
@@ -26,4 +26,4 @@ jobs:
       - name: Get Netlify deployment URL
         uses: voorhoede/get-netlify-url@v1
         with:
-            site-name: test-rcl
+            site-name: gecraft-bsa
