+++ src/components/Book/styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const BookContainer = styled.div`
+  button {
+    margin: 10px 0 0 15px;
+    text-decoration: none;
+    display: inline-block;
+    padding: 8px 16px;
+  }
+
+  button:hover {
+    background-color: #ddd;
+    color: black;
+  }
+`;
