+++ src/components/MenuBar/MenuBar.js
@@ -1,0 +1,54 @@
+import React from 'react';
+
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import InputBase from '@material-ui/core/InputBase';
+
+import SelectLanguage from './SelectLanguage/SelectLanguage';
+
+import SearchIcon from '@material-ui/icons/Search';
+
+import useStyles from './style';
+
+const MenuItems = [
+  { title: 'Bible' },
+  { title: 'OBS' },
+  { title: 'Bible in your language' },
+];
+
+export default function MenuBar() {
+  const classes = useStyles();
+  const Menu = MenuItems.map((MenuItems, index) => (
+    <Typography className={classes.menuItem} variant="h6" noWrap key={index}>
+      {MenuItems.title}
+    </Typography>
+  ));
+
+  return (
+    <div className={classes.root}>
+      <AppBar className={classes.appBar} position="static">
+        <Toolbar>
+          <Typography className={classes.title} variant="h6" noWrap>
+            ROB Resources
+          </Typography>
+          <div className={classes.menu}>{Menu}</div>
+          <div className={classes.search}>
+            <div className={classes.searchIcon}>
+              <SearchIcon />
+            </div>
+            <InputBase
+              placeholder=""
+              classes={{
+                root: classes.inputRoot,
+                input: classes.inputInput,
+              }}
+              inputProps={{ 'aria-label': 'search' }}
+            />
+          </div>
+          <SelectLanguage />
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
