--- src/components/NavigationBar/styled.js
@@ -1,20 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  height: 55px;
-  background-color: black;
-  color: white;
-  display: flex;
-  align-items: center;
-`;
-
-export const Header = styled.div`
-  font-size: 25px;
-`;
-
-export const NavigationLink = styled.a`
-  text-decoration: none;
-  :visited {
-    color: white;
-  }
-`;
