--- src/components/SearchBar/styled.js
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  align-items: center;
-  img {
-    height: 20px;
-  }
-`;
-
-export const SearchInput = styled.input`
-  background-color: white;
-  height: 30px;
-  border-radius: 18px;
-  width: 190px;
-  padding: 0 10px;
-`;
