+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+
+import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
